Fix grammar bugs
We need to include the parser file that LALRPOP generates in our module
tree, so that Rust will compile it. This is what the [`lalrpop_mod`][lm]
macro does.

Note that unlike many other languages (e.g. Java, Python), Rust doesn't
infer the module hierarchy automatically from source files. Instead, the
hierarchy is explicitly defined through the [`mod`][kw] keyword, and is
decoupled from the file system.

More references on modules:

- https://doc.rust-lang.org/reference/items/modules.html
- https://doc.rust-lang.org/rust-by-example/mod.html
- https://www.sheshbabu.com/posts/rust-module-system/

After compiling the generated code, I fixed a few errors in the grammar file
(mostly typos). The last two type errors are a little trickier.

The first one is due to a difference between `&str` and `String`. The Java
analogues would be `String` and `StringBuilder`, or in C++ you have
`std::string_view` and `std::string`. For more information:

- https://blog.thoughtram.io/string-vs-str-in-rust/
- https://stackoverflow.com/questions/24158114/what-are-the-differences-between-rusts-string-and-str
- https://blog.mgattozzi.dev/how-do-i-str-string/

The second is using the implementation of [`std::str::FromStr`][fs] for i32.

[lm]: https://docs.rs/lalrpop-util/latest/lalrpop_util/macro.lalrpop_mod.html
[kw]: https://doc.rust-lang.org/std/keyword.mod.html
[fs]: https://doc.rust-lang.org/std/str/trait.FromStr.html