Remove `do_notation` dependency
This is Rust, not Haskell :P

While Rust doesn't have a full `Monad` trait (because it doesn't have
higher-kinded types), it does essentially special-case the `Option`
and `Result` types as the most common use-cases for monads.

I think there's some work to generalize the question mark operator
(essentially the same as monadic bind) via the [`Try` trait][try],
but not clear on the stabilization timeline.

People have come up with work-arounds for higher kinded types, but
they're so (syntactically) ugly and convoluted that I doubt they'll
see much adoption:

- https://rustyyato.github.io/type/system,type/families/2021/02/15/Type-Families-1.html
- https://hugopeters.me/posts/14/ (https://news.ycombinator.com/item?id=29508790)
- https://www.reddit.com/r/rust/comments/6seaf8/why_are_higher_kinded_types_requested_in_rust/

[try]: https://doc.rust-lang.org/std/ops/trait.Try.html